create a button lifecycle state machine
this makes it easy to know what the current state of the button is. have
we been pressed once? twice? etc.

I'd like to change this from "awaiting..." to "at..." to be a little
clearer.

next we'll add a watcher kinda thing that checks on the button like
500ms after the first press to see what state the button is in and
trigger the appropriate action.