start making an internal/external decorator
I want an API for the connection where someone can pass in an IP/port
pair and get a connection that handles retries and errors and reconnects
all on its own without the caller really needing to worry about it.
thus, a CasetaConnection and an InternalCasetaConnection class. When
CasetaConnection encounters a problem (likely a dead connection or an
auth issue), it will create and swap in a new InternalCasetaConnection
instance and retry. and the caller won't really need to orchestrate
anything.