YAGNI: watch remotes instead of buttons
this is a work in progress. really, we care about remotes activities,
not buttons' activities. if someone were to double-press two buttons on
a remote, we might want to use that to precipitate some other kind of
event. but it's a lot more complicated to do this, so I very well might
revert this commit.

but the work was interesting enough to warrant me putting it into VCS in
case I want to revisit it at some point