start stripping out color/device/etc info
it turns out that trying to figure out the current hue scene is pretty
tricky. and a reasonable but far simpler implementation would be to
start with assuming that the current scene is the first scene in the
configuration, moving on to the next scene, and storing that in a cache
with a ttl of a minute or so. it's not critical that we know the current
scene at any given time, but it is important that we know the current
scene when the user is paging through the scenes (ie not for the first
press of "next scene," but for the second and subsequent presses of
"next scene")

so let's strip out all of the color/device code, which will simplify
configuration and business logic