fix a bug with long presses
for certain buttons, a long press would result in the app seeing an
initial button action of a release. and the app should ignore those, but
it wasn't. so it was enqueueing all of the remote watching machinery and
ultimately that machinery got into a wonky state since it assumed that
first button actions would always be presses.

this change makes the app ignore initial button releases and fail loudly
if one sneaks through the cracks.