allow dispatcher to spawn many tasks
previously, the dispatcher would be beholden to the mpsc receiver
channel queue and make an API request when it saw a new message. But
these requests all happened within the same task, so despite all of the
async machinery, the requests would happen serially in a blocking
fashion. not cool