better turn on behavior
previously, this just sent the "turn on" command to hue, which would
turn on the lights to whatever settings they were previously on. that
doesn't expose the current scene, so if it turned on to the first scene,
we wouldn't necessarily know. and if you pushed the "favorite" button,
we'd move to the "next" scene, which we considered to be the first
configured scene.

this meant that hitting the favorite button in this case wouldn't update
change the scene, so the app would feel like it's not responding.

now, we always turn on to the first scene if we don't know what scene we
most recently had on (i.e. the cache entry for the current scene is
empty).