Fix bottleneck (#2)
* allow dispatcher to spawn many tasks

previously, the dispatcher would be beholden to the mpsc receiver
channel queue and make an API request when it saw a new message. But
these requests all happened within the same task, so despite all of the
async machinery, the requests would happen serially in a blocking
fashion. not cool

* better update values/timing

a user shouldn't have to hold the button for 10 seconds to go to min/max
brightness

* add async mutex for each room in the dispatcher

before this synchronization, the lagginess of the hue API was causing
problems with out-of-order update execution. If a request was sent and
another request was sent and received before the first request
responded, the first request could overwrite the internal state of the
dispatcher

adding some synchronization here lets us run many tasks to update many
rooms, but, for any given room, run the updates serially one at a time
thank u rust for making sure I don't share mutable things


* better turn on behavior

previously, this just sent the "turn on" command to hue, which would
turn on the lights to whatever settings they were previously on. that
doesn't expose the current scene, so if it turned on to the first scene,
we wouldn't necessarily know. and if you pushed the "favorite" button,
we'd move to the "next" scene, which we considered to be the first
configured scene.

this meant that hitting the favorite button in this case wouldn't update
change the scene, so the app would feel like it's not responding.

now, we always turn on to the first scene if we don't know what scene we
most recently had on (i.e. the cache entry for the current scene is
empty).

* fix negative number modulo

in some languages, negative numbers modulo output is positive (i.e. they
wrap around). not so in rust.