start working on a decorator/delegate pattern
I want an external-facing connection manager that delegates to and
manages the lifecycle of an internal connection object. when it notices
the internal conncetion getting mangled, it transparently replaces the
internal connection.

the clients of the external connection manager should be none the wiser
about the internal manager/replaces/etc.