more progress on the decorator
learning lots more about traits, dyn, and lifetimes here. that's neat!

I'm still having trouble figuring out how to spawn the
check-for-connection-liveness loop without the &self escaping out of the
function call. this is a valid error, because we don't know how long the
reference to &self will be valid. but some combination of lifetimes
ought to be able to sort this out. tbd