fixed the bug and added lotsa logging
the bug was that interval::tick returns immediately cuz it's the
first iteration of the interval. so that branch of the tokio::select!
statement always finished first. And since the keepalive message that
was getting sent caused caseta to return a new GNET> line, we were
decoding a bunch of logged in prompt lines that we didn't really need to
hold on to.

now we throw away those GNET lines in the await_message impl for the
delegating connection. and we have a whole bunch of logging