don't automatically restart containers
if there's an issue, let's just replace the container with a new one.

This is important because we sometimes ended up with multiple
containers. the service configuration specifies that we should have one
replica. when we restarted a container in the past, the service manager
would think that we had no replicas up, so it created a new container to
satisfy the required replicas. then the restarting container came back
on line, leading to two replicas running when only one was specified.