turn on room if it's off and any button is pressed
if the lights in a room are off, I can't really see which button I'm
pressing. so currently, I have to feel the remote and feel my way to the
power on button. Things would be a lot simpler if all of the buttons
behaved like the power on button when the room is off.

e.g. do you want to turn the lights on, but you're unable to see the
remote to find the power on button? good news! now all the buttons are
power on buttons!